Implement plugin interface
This adds a plugin interface to kopano-apid to dynamically load api
endpoints from other shared modules.

To support this kopano-apid is no longer static but linked with libc to
support shared module loading.

An example plugin was added to `plugins/example-plugin`.

To configure the location of the plugins files a `--plugins-path`
parameter is added which automatically tries to load and register all
files with `.so` in the name on startup.