Disable obsolete groupware-core v0 API endpoints
This change disables the v0 endpoints for the Kopano Groupwre REST API
by default. To enable it for historic backwards compatibility purposes
the environment variable KOPANO_GC_ENABLE_API_V0 can be set to 1 when
kapi is started. Note that the V0 API does not check authorized scopes
and thus should no longer be enabled in production.