groupware-core: Include socket folder check in retry
When kapid is started it is not guaranteed that the socket folder from
MFR is already there (since MFR might not be started yet). This change
includes the check for the folder in the retry loop so kapid starts up
properly and waits until the folder with its socket appears.