Implement initialization order for plugins
To allow controlled URL routing, it is required to control the order of
the plugins so that a plugin can handle a subset of the URL routing of
another plugin. This change uses the already existing `--plugins`
commandline parameter as order. If given, the plugins are initialized
exatly in the order as given in that commandline parameter value.