Move plugins directly into daemon binary
Due to problems with Go vendoring and plugins which are basically
currently incompatible this change gets rid of the isolated plugin files
and instead directly includes all plugins in the main binary. Since it
was never supported anyways to compile plugins while not compiling the
daemon at the same time, this change is of no functional consequence but
avoids a lot of problems. The plugins concept might be revisited in the
future. For now, anyone building kapi must also build and include all
desired plugins in the build